wayback-restore/requirements.txt
requests==2.32.5
pyyaml==6.0.3
boto3==1.42.55
sqlalchemy==2.0.46
mysqlclient==2.2.8
pytest==9.0.2
jinja2==3.1.6
pylint==4.0.5
validators==0.35.0
pytemplate-reverse==0.0.3
ratelimit==2.2.1
beautifulsoup4==4.14.3
backoff==2.2.1
scrapy==2.14.1
playwright==1.58.0